requirements.txt
ollama>=0.4.4
meshtastic>=2.5.5
pytap2>=2.3.0
python-dotenv>=1.0.1
pyyaml>=6.0.2
bcrypt>=4.0.0
passlib[bcrypt]>=1.7.0
cryptography>=3.4.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
websockets>=12.0
httpx[socks]>=0.27.0
